# Backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AHFUL Users API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/" }
  ],
  "paths": {

    "/AHFULusers/": {
      "get": {
        "summary": "Get all users",
        "tags": ["Users"],
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "object" }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/AHFULusers/{email}": {
      "get": {
        "summary": "Get user by email",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "description": "Email address of the user",
            "schema": { "type": "string", "format": "email" }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string", "example": "6993a3b2b684e1023202a5e9" },
                    "email": { "type": "string", "format": "email", "example": "jane@example.com" },
                    "name": { "type": "string", "example": "Jane Doe" },
                    "password": {
                      "type": "string",
                      "example": "scrypt:32768:8:1$Hp6N4Svxx8qtF6Eg$7b80288755977aeecb541720c2dd77a1d27b85aee23d905b7e4990f8f776d491082e928817269610897c77e82e2acc04fe2c04cd94dce3aebf5ff9ea44673319"
                    },
                    "role": { "type": "integer", "example": 0 }
                  },
                "required": ["_id", "email", "name", "password", "role"]
                }
              }
            }

          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { 
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/AHFULusers/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password", "role"],
                "properties": {
                  "name": { "type": "string", "example": "Jane Doe" },
                  "email": { "type": "string", "format": "email", "example": "jane@example.com" },
                  "password": { "type": "string", "format": "password", "example": "P@ssw0rd!" },
                  "role": { "type": "int", "example": "0" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string", "format": "email" },
                    "message": { "type": "string", "example": "User created successfully" }
                  },
                  "required": ["email", "message"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or registration error",
            "content": {
              "application/json": {
                "schema": { 
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/AHFULusers/login": {
      "post": {
        "summary": "Authenticate a user",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "jane@example.com" },
                  "password": { "type": "string", "format": "password", "example": "P@ssw0rd!" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["email", "message"],
                  "properties": {
                    "email": { "type": "string", "format": "email", "example": "jane@example.com" },
                    "message": { "type": "string", "example": "Login successful" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No data provided or invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": { "type": "string", "example": "No data provided" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": { "type": "string", "example": "Invalid email or password" }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/AHFULworkout/": {
      "get": {
        "summary": "Get all workouts",
        "tags": ["Workout"],
        "responses": {
          "200": {
            "description": "A list of workouts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "object" }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/AHFULworkout/{email}": {
      "get": {
        "summary": "Get workouts by user email",
        "tags": ["Workout"],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "description": "Email address of the user",
            "schema": { "type": "string", "format": "email" }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string", "example": "6993a3b2b684e1023202a5e9" },
                    "email": { "type": "string", "format": "email", "example": "anEmail@email.com" },
                    "title": { "type": "string", "example": "Daily workout" },
                    "gymId": { "type": "integer", "example": 0 },
                    "startTime": { "type": "integer", "example": 0 },
                    "endTime": { "type": "integer", "example": 0 }
                  },
                "required": ["_id", "email", "startTime"]
                }
              }
            }

          },
          "404": {
            "description": "Workouts not found",
            "content": {
              "application/json": {
                "schema": { 
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/AHFULworkout/id/{id}": {
      "get": {
        "summary": "Get workout by id",
        "tags": ["Workout"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "id of the workout",
            "schema": { "type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Workout found",
            
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string", "example": "698d039a6e5117c22dd7771d" }
                  },
                "required": ["_id"]
                }
              }
            }

          },
          "404": {
            "description": "Workout not found",
            "content": {
              "application/json": {
                "schema": { 
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/AHFULworkout/create": {
      "post": {
        "summary": "Create a new workout",
        "tags": ["Workout"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "start_time"],
                "properties": {
                  "title": { "type": "string", "example": "A workout" },
                  "email": { "type": "string", "format": "email", "example": "jane@example.com" },
                  "gymId": { "type": "int", "example": "0" },
                  "startTime": { "type": "int", "example": "0" },
                  "endTime": { "type": "int", "example": "0" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workout created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string", "format": "email" },
                    "message": { "type": "string", "example": "User created successfully" }
                  },
                  "required": ["email", "message"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or creation error",
            "content": {
              "application/json": {
                "schema": { 
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }

  }
}